requirements.txt
aiofiles
aiohttp>=2.2.5
alphabet_detector
apscheduler
asyncio
beautifulsoup4
bing_image_downloader
bleach
bs4
cachetools
countryinfo
CurrencyConverter
dateparser
dnspython
emoji==1.7.0
emoji-country-flag
feedparser
future
geopy
gpytranslate
hachoir
heroku3
html2text
jikanpy
lxml
markdown2
motor
nltk
opencv-python-headless
Pillow
pretty_errors
psutil
psycopg2-binary
PyDictionary
pytz
pymongo
python-arq
pyrate-limiter
pykeyboard
pynewtonmath
pyrogram==2.0.59
python-telegram-bot==12.8
regex
requests
secureme
search_engine_parser
speedtest-cli
spongemock
sqlalchemy==1.4.42
telegram
telegraph
telethon==1.25.4
TruthDarePy
TgCrypto
ujson
wget
wikipedia
youtube_search
youtube_search_python
yt_dlp
zalgo-text
gitpython
